requirements.txt
torch
torchvision
timm
transformers
pillow
faiss
matplotlib